Fix tests and analysis issues
