+++ packages/datadog_inappwebview_tracking/android/src/main/AndroidManifest.xml
@@ -1,3 +1,2 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.datadoghq.flutter.inappwebview_tracking">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
 </manifest>
