+++ packages/datadog_inappwebview_tracking/example/android/.gitignore
@@ -1,8 +1,6 @@
 gradle-wrapper.jar
 /.gradle
 /captures/
-/gradlew
-/gradlew.bat
 /local.properties
 GeneratedPluginRegistrant.java
