--- packages/datadog_inappwebview_tracking/example/android/app/src/main/kotlin/com/datadog/inappwebview_tracking_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.datadog.inappwebview_tracking_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()
