+++ packages/datadog_inappwebview_tracking/example/android/app/src/main/kotlin/com/datadoghq/flutter/inappwebviewtracking/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.datadoghq.flutter.inappwebviewtracking.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()
