+++ packages/datadog_inappwebview_tracking/example/android/build.gradle
@@ -1,3 +1,18 @@
+buildscript {
+    repositories {
+        google()
+        mavenCentral()
+        maven {
+            url "https://plugins.gradle.org/m2/"
+        }
+    }
+
+    dependencies {
+        classpath "org.jlleitschuh.gradle:ktlint-gradle:11.6.0"
+        classpath "io.gitlab.arturbosch.detekt:detekt-gradle-plugin:1.19.0"
+    }
+}
+
 allprojects {
     repositories {
         google()
@@ -11,6 +26,9 @@ subprojects {
 }
 subprojects {
     project.evaluationDependsOn(":app")
+
+    apply from: "${project.rootDir}/buildscripts/ktlint.gradle"
+    apply from: "${project.rootDir}/buildscripts/detekt.gradle"
 }
 
 tasks.register("clean", Delete) {
