+++ packages/datadog_inappwebview_tracking/example/android/buildscripts/detekt.gradle
@@ -1,0 +1,13 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+apply plugin: 'io.gitlab.arturbosch.detekt'
+detekt {
+  config = files("${project.rootDir}/detekt.yml")
+  reports {
+    xml.enabled = true
+  }
+}
