+++ packages/datadog_inappwebview_tracking/example/android/buildscripts/ktlint.gradle
@@ -1,0 +1,11 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+apply plugin: 'org.jlleitschuh.gradle.ktlint'
+ktlint {
+    android = true
+    outputColorName = "RED"
+}
